Add files via upload
Project Task:
The objective of this project is to write a program implementing the generic birthday attack
to find a collision. More specifically, the program runs the birthday attack against the
BadHash44 function.
Birthday Attack
A Birthday attack is a cryptography attack based on the birthday paradox in probability theory.
In probability theory, the birthday paradox means there is a good probability of two randomly
selected people having the same birthday. In the context of cryptography, this concept is
exploited between two or more parties. The attack has a fixed number of pigeonholes or
permutations and the probability of collision increases with an increase in random attack
attempts [1].
Collision search using Birthday Attack
For our project, BadHash44() is a hash function that intentionally returns a small hash output
generated using SHA256. It takes a random 256-bit input message, computes the SHA256
hash value, and returns only the first 44 bits of the hash. We will be finding a collision from
BadHash44 by continuously giving it random input values.
The two tasks that will be done in this project are:
1. Search for collision in BadHash44 by finding two random inputs when given to
BadHash44 as input, outputs the same hash value. When a collision is found, the two
input messages and the corresponding bad hash value are reported.
2. Compute the hashes for the random messages in a table, sort the table on the hash
value, and store them in a “hash.txt” or a “hash.csv” file. The BadHash44 function is
run against all the random messages from the messages list.
Program Description:
For this project, the algorithm is written in Python 3.9.7 and run in the Anaconda Jupyter
Notebook environment.